requirements.txt
asgiref==3.2.10
Django==3.0.8
django-ckeditor==5.9.0
django-environ==0.4.5
django-js-asset==1.2.2
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.2.0
gunicorn==20.1.0
psycopg2==2.9.1